Replaced Google Analytics with SegmentAnalytics.jsx
