+++ package.json
@@ -49,7 +49,6 @@
     "react-cropper": "^1.2.0",
     "react-datepicker": "1.4.1",
     "react-dom": "^17.0.2",
-    "react-ga": "^2.5.7",
     "react-helmet": "^5.2.1",
     "react-helmet-async": "^1.0.2",
     "react-html-parser": "^2.0.2",
