+++ src/components/RequestMarkTraceCompleteButton.jsx
@@ -5,7 +5,6 @@ import TraceService from 'services/TraceService';
 import Trace from 'models/Trace';
 import ErrorPopup from 'components/ErrorPopup';
 import ConversationModal from 'components/ConversationModal';
-import GA from 'lib/GoogleAnalytics';
 import { checkBalance, actionWithLoggedIn } from 'lib/middleware';
 import { Context as Web3Context } from '../contextProviders/Web3Provider';
 import { Context as UserContext } from '../contextProviders/UserProvider';
@@ -69,10 +68,13 @@ const RequestMarkTraceCompleteButton = ({ trace, isAmountEnoughForWithdraw }) =>
                 from: userAddress,
                 proof,
                 onTxHash: txUrl => {
-                  GA.trackEvent({
+                  window.analytics.track('Trace Marked Complete', {
                     category: 'Trace',
                     action: 'marked complete',
                     label: trace._id,
+                    title: trace.title,
+                    userAddress: currentUser.address,
+                    txUrl,
                   });
 
                   React.toast.info(
